requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib
fastapi-mail
jinja2
httpx
google-auth
google-auth-oauthlib
python-jose
itsdangerous
python-multipart
